Prefer libopenblaso.so (openblas-openmp)
Fedora and RHEL-based systems provide a plethoria of OpenBLAS builds.
The default `libopenblas.so` is provided by single-threaded
`openblas-serial` package. The OpenMP-aware variant is `libopenblaso.so`
from the `openblas-openmp` package. Other variants include `libopenblasp.so`
(pthread), ILP64 variant like `libopenblaso64.so` and ILP64 with suffix like
`libopenblaso64_.so`.

`FindBLAS` and `FindOpenBLAS` now check for and prefer the OpenMP
variant `openblaso`.

Signed-off-by: Christian Heimes <cheimes@redhat.com>